Extract Remote capabilities, products and receivers to their own components
